more WIP on making tanks collide-able with missiles
This change cleans up the setting of a tank's cx and cy member
variables by moving them all into member functions (of Tank).  The idea
is to consolidate functionality that affects the collide able contents
of a tank's body, such that it can be updated as needed, with greater
ease.